requirements.txt
flask
markdown
python-slugify
